setup mpi_read_map test for runtests
